Added a Queue for buzzing during Tossup
I'm gonna try doing this stuff more often, but on topic. I've started to test with more people and this was the
biggest problem of having to buzz as quickly as possible after someone else finished their buzz. I moved buzzing
to a inline function and promisified. Currently the queue is reduced through a recursive function but in the
future it should be dynamic then callbacks. I also added custom events that allow for two tossups having at
different times.

Changes:
  - Buzz Queue
  - Dynamic Buzz Event Calls
  - Fixed the Guild Adding DM
  - Removed random Console debugs
  - Typos!