Added Power on voice procotoring
This was an easy fix that took a lot of thought. There were many solutions but I settled for the easy way out, spliting up the question into to two
audio files (before and after) to check for power. Other solutions that could be researched are speech recognition, but it seemed to complicated for
the use case. The current solution is also pretty annoying with a big gap between the audio files. I don't have any solutions but for now, voice
proctoring is now an option and text based is the default. I also moved single commands without prefixes such as buzz to their Collection in time for
tourneys.

Changes:
  - Power in voice procotoring
  - Text based proctoring is now default
  - Separate Collection for commands such as buzz
  - Solved the stupid Warning about recreating QueryFiles